# src/Web/Vinva.Gastronomy.WebClient/src/api/gastronomy-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gastronomy",
    "description": "Gastronomy API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Categories/Create": {
      "post": {
        "tags": [
          "Category"
        ],
        "operationId": "CategoryCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCategoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Categories/Remove": {
      "delete": {
        "tags": [
          "Category"
        ],
        "operationId": "CategoryRemove",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveCategoryCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryDtoType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "CreateCategoryCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CategoryDtoType"
          }
        },
        "additionalProperties": false
      },
      "CreateCategoryResponse": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "RemoveCategoryCommand": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "description": "Cookie based configuration",
        "name": "accessToken",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "cookieAuth": [ ]
    }
  ]
}